finally read task state segment from memory

+++ intermediate-x86/kernel/hexdump_gdt_memory/hexdump_gdt_memory.c
@@ -7,17 +7,17 @@
 
 static int __init my_module_init(void) {
     struct desc_ptr gdt;
-    unsigned long long *ptr;
-    unsigned long long *end;
+    unsigned long *ptr;
+    unsigned long *end;
     int i = 0;
 
     native_store_gdt(&gdt);
     printk(KERN_ALERT "GDT descriptor address: %#lx\n", gdt.address);
     printk(KERN_ALERT "GDT descriptor size: %#x\n", gdt.size);
 
-    end = (long long *) gdt.address + gdt.size;
+    end = (long *) gdt.address + gdt.size;
 
-    for (ptr = (long long *) gdt.address; ptr < end; ptr++) {
+    for (ptr = (long *) gdt.address; ptr < end; ptr++) {
         printk(KERN_ALERT "index: %d -> %#lx\n", i, *ptr);
         i++;
     }
